+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed a bug where variants’ `promotionalPrices` weren’t showing as currency values on variant indexes.
 - Fixed a PHP error that could occur when sending an order email. ([#3596](https://github.com/craftcms/commerce/issues/3596))
+- Fixed a bug where dimension fields were not displaying values in the correct formatting locale. ([#3636](https://github.com/craftcms/commerce/issues/3636))
 
 ## 5.0.16.2 - 2024-08-16